Refactor folder structure for domain screens
