+++ lib/screens/domains/domains.dart
@@ -11,8 +11,8 @@ import 'package:pi_hole_client/models/server.dart';
 import 'package:pi_hole_client/providers/app_config_provider.dart';
 import 'package:pi_hole_client/providers/domains_list_provider.dart';
 import 'package:pi_hole_client/providers/servers_provider.dart';
-import 'package:pi_hole_client/screens/domains/domain_details_screen.dart';
-import 'package:pi_hole_client/screens/domains/domains_list.dart';
+import 'package:pi_hole_client/screens/domains/widgets/domain_details_screen.dart';
+import 'package:pi_hole_client/screens/domains/widgets/domains_list.dart';
 import 'package:provider/provider.dart';
 
 class DomainLists extends StatelessWidget {
