+++ lib/screens/domains/widgets/domain_details_screen.dart
@@ -5,7 +5,7 @@ import 'package:pi_hole_client/functions/conversions.dart';
 import 'package:pi_hole_client/functions/format.dart';
 import 'package:pi_hole_client/l10n/generated/app_localizations.dart';
 import 'package:pi_hole_client/models/domain.dart';
-import 'package:pi_hole_client/screens/domains/domain_comment_modal.dart';
+import 'package:pi_hole_client/screens/domains/widgets/domain_comment_modal.dart';
 import 'package:pi_hole_client/widgets/custom_list_tile.dart';
 import 'package:pi_hole_client/widgets/delete_modal.dart';
