+++ lib/screens/domains/widgets/domains_list.dart
@@ -11,9 +11,9 @@ import 'package:pi_hole_client/models/gateways.dart';
 import 'package:pi_hole_client/providers/app_config_provider.dart';
 import 'package:pi_hole_client/providers/domains_list_provider.dart';
 import 'package:pi_hole_client/providers/servers_provider.dart';
-import 'package:pi_hole_client/screens/domains/add_domain_modal.dart';
-import 'package:pi_hole_client/screens/domains/domain_details_screen.dart';
-import 'package:pi_hole_client/screens/domains/domain_tile.dart';
+import 'package:pi_hole_client/screens/domains/widgets/add_domain_modal.dart';
+import 'package:pi_hole_client/screens/domains/widgets/domain_details_screen.dart';
+import 'package:pi_hole_client/screens/domains/widgets/domain_tile.dart';
 import 'package:pi_hole_client/widgets/tab_content_list.dart';
 import 'package:provider/provider.dart';
