+++ test/full_coverage_test.dart
@@ -95,12 +95,12 @@ import 'package:pi_hole_client/screens/app_logs/app_logs.dart';
 import 'package:pi_hole_client/screens/common/empty_data_screen.dart';
 import 'package:pi_hole_client/screens/common/error_message_screen.dart';
 import 'package:pi_hole_client/screens/common/pi_hole_v5_not_supported_screen.dart';
-import 'package:pi_hole_client/screens/domains/add_domain_modal.dart';
-import 'package:pi_hole_client/screens/domains/domain_comment_modal.dart';
-import 'package:pi_hole_client/screens/domains/domain_details_screen.dart';
-import 'package:pi_hole_client/screens/domains/domain_tile.dart';
 import 'package:pi_hole_client/screens/domains/domains.dart';
-import 'package:pi_hole_client/screens/domains/domains_list.dart';
+import 'package:pi_hole_client/screens/domains/widgets/add_domain_modal.dart';
+import 'package:pi_hole_client/screens/domains/widgets/domain_comment_modal.dart';
+import 'package:pi_hole_client/screens/domains/widgets/domain_details_screen.dart';
+import 'package:pi_hole_client/screens/domains/widgets/domain_tile.dart';
+import 'package:pi_hole_client/screens/domains/widgets/domains_list.dart';
 import 'package:pi_hole_client/screens/home/disable_modal.dart';
 import 'package:pi_hole_client/screens/home/home.dart';
 import 'package:pi_hole_client/screens/home/home_appbar.dart';
